+++ src/lib.rs
@@ -110,7 +110,7 @@
 //! - `wl-copy`: implements `wl-copy` from [wl-clipboard](https://github.com/bugaevc/wl-clipboard).
 //! - `wl-clip`: a Wayland version of `xclip`.
 
-#![doc(html_root_url = "https://docs.rs/wl-clipboard-rs/0.2.0")]
+#![doc(html_root_url = "https://docs.rs/wl-clipboard-rs/0.3.0")]
 #![deny(unsafe_code)]
 
 mod common;