+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - **Breaking** Moved `ClipboardType` into `copy::` and `paste::`.
+- Added `copy::ClipboardType::Both` for operating both clipboards at once.
 
 ## v0.2.0 (17th Feb 2019)